backend/requirements.txt
Flask
flask-cors
paho-mqtt
python-dotenv